Move the view macros
